requirements.txt
#Python==2.7
django==1.6
pillow
mysql-python
South
django-compressor==1.3
django-appconf==0.6
easy-thumbnails
django-mptt>=0.2.1
django-polymorphic>=0.2
simplejson>=3.3.1
django-haystack>=2.1.0
whoosh>=2.5.6
tweepy
django-suit
django-chunks
jsmin